2_6_9_final
OK

+++ EX_1_2_1/ViewController.m
@@ -144,7 +144,7 @@
     
     NSIndexPath *indexPath = [self.table indexPathForCell:sender];
     
-    Product *selectedProduct = [[Catalog defaultCatalog] productAt:indexPath.row];
+    Product *selectedProduct = [[Catalog defaultCatalog] productAt:(int)indexPath.row];
     
     detailVC.productCode = selectedProduct.code;
